requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
alpaca-py>=0.43.0
pandas>=2.2.0
numpy>=1.26.0,<2.0.0
ta==0.11.0
requests==2.31.0
python-dotenv==1.0.0
httpx>=0.26.0
pytest==7.4.3
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# AI Agent System
anthropic>=0.40.0
aiosqlite>=0.19.0
PyJWT>=2.8.0

# Authentication (Supabase)
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0
websockets>=14.0

# PostgreSQL (production database)
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Security & Rate Limiting
slowapi>=0.1.9
